Build testNameFilter using describes
The testNameFilter was originally only using testId.  This caused it to filter to any tests that contained the target test in their name.  When filtering on a describe block, this meant that no filtering was applied, and the entire file was always run.

This commit updates the testNameFilter to include ancestor describe blocks in the name so that the filter given to Jest is more specific.  This should improve performance in large test files.